Fixed integer overflow in length check
If the start sector was near Int64.int_max, adding on the length of the
buffer would cause it to wrap and incorrectly pass the test.

This is mainly just paranoia, because the underlying device should
reject the invalid start sector, but it doesn't hurt to be careful...
especially if it uses Int64.to_int to get an integer sector number
(Int64.to_int silently wraps!).